refactor: extract rlib stub compilation into the separate module
